ADD | added lazy alpha contract, updated batch buy method, refactored contract names #LASO-0
